+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Current (main branch)
 
+## 1.4.0
+
 * Fixed updating linked app components.
 * Fixed errors during app component update not being reported.
 * New `lvl system sshconfig` command to add system names to your SSH config. This allows easier access via non-lvl commands such as `rsync`.